requirements/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_pytorch_select=0.2=gpu_0
apex=0.1=pypi_0
blas=1.0=mkl
bzip2=1.0.8=h7b6447c_0
ca-certificates=2019.9.11=hecc5488_0
cairo=1.14.12=h8948797_3
certifi=2019.9.11=py37_0
cffi=1.13.2=py37h2e261b9_0
chardet=3.0.4=pypi_0
cudatoolkit=10.0.130=0
cudnn=7.6.4=cuda10.0_0
cycler=0.10.0=py37_0
dbus=1.13.12=h746ee38_0
easydict=1.9=py_0
expat=2.2.6=he6710b0_0
ffmpeg=4.0=hcdf2ecd_0
fontconfig=2.13.0=h9420a91_0
freeglut=3.0.0=hf484d3e_5
freetype=2.9.1=h8a8886c_1
glib=2.63.1=h5a9c865_0
graphite2=1.3.13=h23475e2_0
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb453b48_1
harfbuzz=1.8.8=hffaf4a1_0
hdf5=1.10.2=hba1933b_1
icu=58.2=h9c2bf20_1
idna=2.8=pypi_0
inplace-abn=1.0.7=pypi_0
intel-openmp=2019.4=243
jasper=2.0.14=h07fcdf6_1
jpeg=9b=h024ee3a_2
jsonpatch=1.24=pypi_0
jsonpointer=2.0=pypi_0
kiwisolver=1.1.0=py37he6710b0_0
libedit=3.1.20181209=hc058e9b_0
libffi=3.2.1=hd88cf55_4
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libglu=9.0.0=hf484d3e_1
libopencv=3.4.2=hb342d67_1
libopus=1.3=h7b6447c_0
libpng=1.6.37=hbc83047_0
libprotobuf=3.10.1=h8b12597_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_0
libuuid=1.0.3=h1bed415_2
libvpx=1.7.0=h439df22_0
libxcb=1.13=h1bed415_1
libxml2=2.9.9=hea5a465_1
matplotlib=3.1.1=py37h5429711_0
mkl=2019.4=243
mkl-service=2.3.0=py37he904b0f_0
mkl_fft=1.0.15=py37ha843d7b_0
mkl_random=1.1.0=py37hd6b4f25_0
ncurses=6.1=he6710b0_1
ninja=1.9.0=py37hfd86e86_0
numpy=1.17.3=py37hd14ec0e_0
numpy-base=1.17.3=py37hde5b4d6_0
olefile=0.46=py37_0
opencv=3.4.2=py37h6fd60c2_1
openssl=1.1.1d=h516909a_0
pandas=0.25.1=py37he6710b0_0
pcre=8.43=he6710b0_0
pillow=6.2.1=py37h34e0f95_0
pip=19.3.1=py37_0
pixman=0.38.0=h7b6447c_0
protobuf=3.10.1=py37he1b5a44_0
py-opencv=3.4.2=py37hb342d67_1
pycparser=2.19=py37_0
pyparsing=2.4.5=py_0
pyqt=5.9.2=py37h05f1152_2
python=3.7.4=h265db76_1
python-dateutil=2.8.1=py_0
pytorch=1.2.0=cuda100py37h938c94c_0
pytz=2019.3=py_0
pyyaml=5.1.2=py37h7b6447c_0
pyzmq=18.1.0=pypi_0
qt=5.9.7=h5867ecd_1
readline=7.0=h7b6447c_5
requests=2.22.0=pypi_0
scipy=1.3.1=py37h7c811a0_0
setuptools=41.6.0=py37_0
sip=4.19.8=py37hf484d3e_0
six=1.13.0=py37_0
sqlite=3.30.1=h7b6447c_0
tensorboardx=1.9=py_0
tk=8.6.8=hbc83047_0
torchfile=0.1.0=pypi_0
torchprof=0.3.1=pypi_0
torchsummary=1.5.1=pypi_0
torchvision=0.4.0=cuda100py37hecfc37a_0
tornado=6.0.3=py37h7b6447c_0
tqdm=4.36.1=py_0
urllib3=1.25.6=pypi_0
visdom=0.1.8.9=pypi_0
websocket-client=0.56.0=pypi_0
wheel=0.33.6=py37_0
xz=5.2.4=h14c3975_4
yaml=0.1.7=h96e3832_1
zlib=1.2.11=h7b6447c_3
zstd=1.3.7=h0b5b093_0
